Update tabstitlecontrol.css
Sorry about that, using a fixed width of 902px is not suitable in this case, especially considering that the size of the breadcrumb items can vary - Ideally, it would be preferable to include some padding on the left and right sides of the breadcrumbs container. However, for the time being, the current approach will suffice.